Add extra CLI options.
This commit allows the user to be able to pull several functions at once
using the same CodeDB instance. This is much more efficient than running
code-extract several times.

Also, external decls are not included in the database by default
anymore. This saves memory as we don't expect users to request
extracting externals in the typical case.